Add rubocop, fasterer, reek and Appraisal
