+++ .github/workflows/specs.yaml
@@ -1,0 +1,23 @@
+name: Specs
+on: [pull_request]
+jobs:
+  appraisal:
+    name: appraisal
+    runs-on: ubuntu-latest
+    strategy:
+       fail-fast: false
+       matrix:
+         os: [ubuntu]
+         # We cannot add ruby 3.0 until Rails 5.2 is dropped.
+         # https://github.com/rails/rails/issues/41074#issuecomment-759188945
+         ruby: [2.7] # , 3.0, 3.1
+    steps:
+      - name: Check out code.
+        uses: actions/checkout@v1
+      - uses: ruby/setup-ruby@v1
+        with:
+          ruby-version: ${{ matrix.ruby }}
+      - run: bundle check || bundle install
+      - run: bundle exec appraisal install
+      - name: appraisal
+        run: bundle exec appraisal rspec
