+++ .reek.yml
@@ -1,0 +1,119 @@
+---
+detectors:
+  Attribute:
+    enabled: true
+    exclude:
+    - Que::Scheduler
+  BooleanParameter:
+    enabled: false
+    exclude: []
+  ClassVariable:
+    enabled: true
+    exclude: []
+  ControlParameter:
+    enabled: true
+    exclude: []
+  DataClump:
+    enabled: true
+    exclude: []
+    max_copies: 3
+    min_clump_size: 3
+  DuplicateMethodCall:
+    enabled: true
+    max_calls: 3
+    allow_calls: []
+    exclude:
+    - spec # Specs contain a lot of duplication
+    - expect # Specs contain a lot of duplication
+  FeatureEnvy:
+    enabled: false
+  IrresponsibleModule:
+    enabled: false
+    exclude: []
+  LongParameterList:
+    enabled: true
+    exclude: []
+    max_params: 4
+    overrides:
+      initialize:
+        max_params: 5
+  LongYieldList:
+    enabled: true
+    exclude: []
+    max_params: 3
+  MissingSafeMethod:
+    enabled: true
+    exclude: []
+  ModuleInitialize:
+    enabled: true
+    exclude: []
+  NestedIterators:
+    enabled: true
+    exclude: []
+    max_allowed_nesting: 2
+    ignore_iterators:
+    - tap
+  NilCheck:
+    enabled: false
+    exclude: []
+  RepeatedConditional:
+    enabled: true
+    exclude: []
+    max_ifs: 2
+  TooManyInstanceVariables:
+    enabled: true
+    exclude: []
+    max_instance_variables: 6
+  TooManyMethods:
+    enabled: true
+    exclude: []
+    max_methods: 15
+  TooManyStatements:
+    enabled: true
+    exclude:
+    - initialize
+    max_statements: 15
+  TooManyConstants:
+    enabled: false
+  UncommunicativeMethodName:
+    enabled: true
+    exclude: []
+    reject:
+    - /^[a-z]$/
+    - /[A-Z]/
+    accept: []
+  UncommunicativeModuleName:
+    enabled: true
+    exclude: []
+    reject:
+    - /^.$/
+    - /[0-9]$/
+  UncommunicativeParameterName:
+    enabled: true
+    exclude: []
+    reject:
+    - /^.$/
+    - /[0-9]$/
+    - /[A-Z]/
+    - /^_/
+    accept: []
+  UncommunicativeVariableName:
+    enabled: true
+    exclude: []
+    reject:
+    - /[0-9]$/
+    - /[A-Z]/
+    accept:
+    - _
+  UnusedParameters:
+    enabled: false # This is covered by Rubocop's Lint/UnusedMethodArgument
+    exclude: []
+  UtilityFunction:
+    enabled: true
+    public_methods_only: true
+    exclude:
+    - job_args_from_db_row
+
+exclude_paths:
+- vendor
+- gemfiles
