+++ .rspec
@@ -1,3 +1,2 @@
+--format documentation
 --color
---order random
---require spec_helper
