+++ Appraisals
@@ -1,0 +1,15 @@
+appraise "rails-5-2" do
+  gem "rails", "~> 5.2"
+end
+
+appraise "rails-6" do
+  gem "rails", "~> 6.0.3"
+end
+
+appraise "rails-6-1" do
+  gem "rails", "~> 6.1.0"
+end
+
+appraise "rails-7" do
+  gem "rails", "~> 7.0"
+end
