+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## 1.2.0 (2022-10-11)
+## 1.4.0 (2022-10-14)
 
-* First release. A hard fork of Figaro (with thanks to Steve Richert!). Aliased main module and
-  retained latest published version for drop-in compatibility, but no functional changes.
+- Add rubocop, fasterer, reek and Appraisal [#1](https://github.com/hlascelles/figjam/pull/1)
+
+## 1.3.0 (2022-11-10)
+
+- Aliased main module Figaro to Figjam to allow easier migration. 
+
+## 1.2.0 (2022-11-10)
+
+- First release. A hard fork of Figaro (with thanks to Steve Richert!). Retained latest published
+  version for drop-in compatibility, but no functional changes.
