+++ Gemfile
@@ -1,11 +1,3 @@
 source "https://rubygems.org"
 
 gemspec
-
-gem "rails", ">= 5.2.0", "< 6.1"
-
-group :test do
-  gem "aruba", "~> 1.0"
-  gem "rspec", "~> 3.9"
-  gem "sqlite3"
-end
