+++ appraisal.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+bundle exec appraisal rspec "$@"
