+++ gemfiles/rails_5_2.gemfile
@@ -1,0 +1,7 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "rails", "~> 5.2"
+
+gemspec path: "../"
