+++ gemfiles/rails_6.gemfile
@@ -1,0 +1,7 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "rails", "~> 6.0.3"
+
+gemspec path: "../"
