+++ gemfiles/rails_6.gemfile.lock
@@ -1,5 +1,5 @@
 PATH
-  remote: .
+  remote: ..
   specs:
     figjam (1.3.0)
       thor (>= 0.14.0, < 2)
@@ -66,11 +66,11 @@ GEM
       bundler
       rake
       thor (>= 0.14.0)
-    aruba (1.1.2)
+    aruba (2.1.0)
       bundler (>= 1.17, < 3.0)
       childprocess (>= 2.0, < 5.0)
       contracts (>= 0.16.0, < 0.18.0)
-      cucumber (>= 2.4, < 7.0)
+      cucumber (>= 4.0, < 9.0)
       rspec-expectations (~> 3.4)
       thor (~> 1.0)
     ast (2.4.2)
@@ -82,38 +82,30 @@ GEM
     concurrent-ruby (1.1.10)
     contracts (0.16.1)
     crass (1.0.6)
-    cucumber (6.1.0)
+    cucumber (8.0.0)
       builder (~> 3.2, >= 3.2.4)
-      cucumber-core (~> 9.0, >= 9.0.1)
-      cucumber-create-meta (~> 4.0, >= 4.0.0)
-      cucumber-cucumber-expressions (~> 12.1, >= 12.1.1)
-      cucumber-gherkin (~> 18.1, >= 18.1.0)
-      cucumber-html-formatter (~> 13.0, >= 13.0.0)
-      cucumber-messages (~> 15.0, >= 15.0.0)
-      cucumber-wire (~> 5.0, >= 5.0.1)
-      diff-lcs (~> 1.4, >= 1.4.4)
-      mime-types (~> 3.3, >= 3.3.1)
-      multi_test (~> 0.1, >= 0.1.2)
+      cucumber-ci-environment (~> 9.0, >= 9.0.4)
+      cucumber-core (~> 11.0, >= 11.0.0)
+      cucumber-cucumber-expressions (~> 15.1, >= 15.1.1)
+      cucumber-gherkin (~> 23.0, >= 23.0.1)
+      cucumber-html-formatter (~> 19.1, >= 19.1.0)
+      cucumber-messages (~> 18.0, >= 18.0.0)
+      diff-lcs (~> 1.5, >= 1.5.0)
+      mime-types (~> 3.4, >= 3.4.1)
+      multi_test (~> 1.1, >= 1.1.0)
       sys-uname (~> 1.2, >= 1.2.2)
-    cucumber-core (9.0.1)
-      cucumber-gherkin (~> 18.1, >= 18.1.0)
-      cucumber-messages (~> 15.0, >= 15.0.0)
-      cucumber-tag-expressions (~> 3.0, >= 3.0.1)
-    cucumber-create-meta (4.0.0)
-      cucumber-messages (~> 15.0, >= 15.0.0)
-      sys-uname (~> 1.2, >= 1.2.2)
-    cucumber-cucumber-expressions (12.1.3)
-    cucumber-gherkin (18.1.1)
-      cucumber-messages (~> 15.0, >= 15.0.0)
-    cucumber-html-formatter (13.0.0)
-      cucumber-messages (~> 15.0, >= 15.0.0)
-    cucumber-messages (15.0.0)
-      protobuf-cucumber (~> 3.10, >= 3.10.8)
-    cucumber-tag-expressions (3.0.1)
-    cucumber-wire (5.0.1)
-      cucumber-core (~> 9.0, >= 9.0.1)
-      cucumber-cucumber-expressions (~> 12.1, >= 12.1.1)
-      cucumber-messages (~> 15.0, >= 15.0.0)
+    cucumber-ci-environment (9.1.0)
+    cucumber-core (11.0.0)
+      cucumber-gherkin (~> 23.0, >= 23.0.1)
+      cucumber-messages (~> 18.0, >= 18.0.0)
+      cucumber-tag-expressions (~> 4.1, >= 4.1.0)
+    cucumber-cucumber-expressions (15.2.0)
+    cucumber-gherkin (23.0.1)
+      cucumber-messages (~> 18.0, >= 18.0.0)
+    cucumber-html-formatter (19.2.0)
+      cucumber-messages (~> 18.0, >= 18.0.0)
+    cucumber-messages (18.0.0)
+    cucumber-tag-expressions (4.1.0)
     diff-lcs (1.5.0)
     erubi (1.11.0)
     fasterer (0.10.0)
@@ -132,24 +124,18 @@ GEM
       mini_mime (>= 0.1.1)
     marcel (1.0.2)
     method_source (1.0.0)
-    middleware (0.1.0)
     mime-types (3.4.1)
       mime-types-data (~> 3.2015)
     mime-types-data (3.2022.0105)
     mini_mime (1.1.2)
     minitest (5.16.3)
-    multi_test (0.1.2)
+    multi_test (1.1.0)
     nio4r (2.5.8)
     nokogiri (1.13.9-x86_64-linux)
       racc (~> 1.4)
     parallel (1.22.1)
     parser (3.1.2.1)
       ast (~> 2.4.1)
-    protobuf-cucumber (3.10.8)
-      activesupport (>= 3.2)
-      middleware
-      thor
-      thread_safe
     pry (0.14.1)
       coderay (~> 1.1)
       method_source (~> 1.0)
@@ -203,19 +189,19 @@ GEM
       diff-lcs (>= 1.2.0, < 2.0)
       rspec-support (~> 3.12.0)
     rspec-support (3.12.0)
-    rubocop (1.36.0)
+    rubocop (1.39.0)
       json (~> 2.3)
       parallel (~> 1.10)
       parser (>= 3.1.2.1)
       rainbow (>= 2.2.2, < 4.0)
       regexp_parser (>= 1.8, < 3.0)
       rexml (>= 3.2.5, < 4.0)
-      rubocop-ast (>= 1.20.1, < 2.0)
+      rubocop-ast (>= 1.23.0, < 2.0)
       ruby-progressbar (~> 1.7)
       unicode-display_width (>= 1.4.0, < 3.0)
     rubocop-ast (1.23.0)
       parser (>= 3.1.1.0)
-    rubocop-rspec (2.13.2)
+    rubocop-rspec (2.15.0)
       rubocop (~> 1.33)
     ruby-progressbar (1.11.0)
     ruby_parser (3.19.1)
@@ -239,7 +225,7 @@ GEM
     websocket-driver (0.7.5)
       websocket-extensions (>= 0.1.0)
     websocket-extensions (0.1.5)
-    zeitwerk (2.6.4)
+    zeitwerk (2.6.6)
 
 PLATFORMS
   x86_64-linux
@@ -251,7 +237,7 @@ DEPENDENCIES
   fasterer
   figjam!
   pry-byebug
-  rails
+  rails (~> 6.0.3)
   rake
   rspec
   rubocop
