+++ gemfiles/rails_7.gemfile
@@ -1,0 +1,7 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "rails", "~> 7.0"
+
+gemspec path: "../"
