+++ lib/figjam/version.rb
@@ -1,3 +1,3 @@
 module Figjam
-  VERSION = "1.3.0"
+  VERSION = "1.4.0"
 end
