+++ spec/figjam/rails/application_spec.rb
